Refs #19648: rename xtypes namespace
