+++ include/fastdds/dds/xtypes/type_representation/TypeObject.hpp
@@ -19,12 +19,12 @@
  * This file was generated by the tool fastddsgen.
  */
 
-#ifndef _FAST_DDS_GENERATED_EPROSIMA_FASTDDS_DDS_XTYPES1_3_TYPEOBJECT_H_
-#define _FAST_DDS_GENERATED_EPROSIMA_FASTDDS_DDS_XTYPES1_3_TYPEOBJECT_H_
+#ifndef _FASTDDS_DDS_XTYPES_TYPE_REPRESENTATION_TYPEOBJECT_HPP_
+#define _FASTDDS_DDS_XTYPES_TYPE_REPRESENTATION_TYPEOBJECT_HPP_
 
 #include <fastdds/dds/xtypes/type_representation/detail/dds_xtypes_typeobject.hpp>
 #include <fastdds/dds/xtypes/type_representation/detail/dds_xtypes_typeobjectCdrAux.hpp>
 #include <fastdds/dds/xtypes/type_representation/detail/dds_xtypes_typeobjectPubSubTypes.h>
 
-#endif // _FAST_DDS_GENERATED_EPROSIMA_FASTDDS_DDS_XTYPES1_3_TYPEOBJECT_H_
+#endif // _FASTDDS_DDS_XTYPES_TYPE_REPRESENTATION_TYPEOBJECT_HPP_
